functioning simulator fiddled with some codelines in the neural network, mainly adding weight conversion to outputs. Added the possibility to inherit DNA from a predecessing pixie, in which case no new DNA gets generated. the executeMove-function now actually converts movement urges (moveX, moveY) into movement on the grid. GIFs can now be created if wanted.
A lot of simulator functions added:
eachSimStep, which computes the actions of every pixie during each simStep.
spawnPixie, which instantiates a single new Pixie, optionally with inherited DNA and color.
newGeneration, which generates a whole new world with a new generation of pixies, also optionally (mainly) with inherited genes.
mutateGenes, which alters single bits in the genome with a small chance (currently 0).
saveMetaGenome, a function which stores the DNA of every pixie alive in a textfile
readMetaGenome, which can in turn read a metagenome textfile and convert it back to a genomes list to be fed into newGeneration
simulateGenerations, the main function which initiates the whole simulation from start to finish.