new neurons, selection criteria, bugfixes new neurons include OnOff, setOnOff, nextFood, eatFood. new selection critarium is killLowEnergy new pixie variable isOn (bool) added
a lot of new neurons added to the dictionaries. Worked like a charm with only minor bugfixes needed!

the abs()-function in direction-independent actionNeurons was removed (3 instances).

MOST IMPORTANTLY the queueForMove-mechanism is deactivated. Every executeMove()-
function is immediately called for each pixie and the world is updated. This requires more
calculation steps per simstep/generation, but prevents pixies occupying the same gridcell.

Selection for food search was tested and worked well!