Muller Plot tweaks & a lot of aesthetic changes First of all, the mutateGenes-function has never worked previously without us noticing it. This was fixed, so that mutation now exists.
Concerning the muller plot there have also been some changes:
- the option of showing the "real" color of the pixie is not toggleable
- the list of genomes that are plotted are now sorted alphabetically, which means
that mutated lineages appear next to their mother lineage in the plot (and not at the top)

Also, the mechanism in which color is inherited was changed:
If the genome of a inheriting pixie mutates, its color is also mutated in a way in which
it varies slightly from the original color. This enables us to distinguish mutated lineages
on the grid as well as in the muller plot while still recongnizing them as kin.
This was possible due to 3 new "convenience functions" that chatGPT wrote, to be found in the
new CONVENIENCE FUNCTIONS section: hex_to_rbg, rgb_to_hex and generate_similar_color.